reorganization js components  without code modification
