+++ site/src/components/home/fragment/Slideshow.js
