+++ site/src/components/pages_object/ActivityDetail.js
