+++ site/src/components/pages_object/NewsDetail.js
