+++ site/src/components/pages_object/PeopleDetail.js
