+++ site/src/components/pages_object/ProductionDetail.js
