+++ site/src/components/pages_object/fragment/ActivitesAssociees.js
